pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "phasma"
version = "0.5.0"
description = "A modern PhantomJS driver for Python with bundled binaries, providing headless browser automation, web page rendering, and JavaScript execution."
readme = "README.md"
requires-python = ">=3.4"
license = {file = "LICENSE"}
keywords = ["phantomjs", "headless-browser", "web-automation", "javascript", "web-scraping"]
authors = [
  { name = "Mohammad Raziei", email = "mohammadraziei1375@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Microsoft :: Windows",
  "Topic :: Internet :: WWW/HTTP :: Browsers",
  "Topic :: Software Development :: Testing",
  "Topic :: Software Development :: Libraries :: Python Modules",
]


[project.optional-dependencies]
dev = [
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Source = "https://github.com/mohammadraziei/phasma"
Issues = "https://github.com/mohammadraziei/phasma/issues"
Documentation = "https://mohammadraziei.github.io/phasma" 



[tool.black]
target-version = ["py37"]
line-length = 120
skip-string-normalization = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
    # "error",
    "ignore:(ast.Str|Attribute s|ast.NameConstant|ast.Num) is deprecated:DeprecationWarning:_pytest",  # Python 3.12
]
testpaths = ["tests"]
pythonpath = [
  "."
]


[tool.ruff]
target-version = "py37"
line-length = 120

[tool.ruff.lint]  # Ruff linting configuration
select = [        # List of lint rule codes to enable
  "A",            # Built-in rules for potential errors
  "ARG",          # Argument-related checks
  "B",            # Bugbear: likely bugs and design problems
  "C",            # Complexity checks
  "DTZ",          # Datetime-related issues
  "E", "EM",      # pycodestyle errors, error messages
  "F", "FBT",     # Pyflakes, boolean trap checks
  "I", "ICN", "ISC", # Import sorting, conventions, string concatenation
  "N",            # Naming conventions
  "PLC", "PLE", "PLR", "PLW", # Pylint checks (convention, error, refactor, warning)
  "Q",            # Quotes
  "RUF",          # Ruff-specific rules
  "S",            # Security issues
  "T", "TID",     # Type annotations, tidy imports
  "UP",           # pyupgrade: modernize code
  "W",            # pycodestyle warnings
  "YTT",          # sys.version checks
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
  # Ignore some unsecure rules:
  "S301",
  # Ignore relative imports
  "TID252"
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.lint.isort]
known-first-party = ["phasma"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252", "E501", "F401"]
"__init__.py" = ["F401"]
# Allow print statements and unused imports in example files
"examples/*.py" = ["T201"]

[tool.coverage.run]
source_pkgs = ["phasma"]
branch = true
parallel = true
omit = [
  "phasma/__about__.py",
]


[tool.coverage.paths]
phasma = ["./phasma"]
tests = ["tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "__version__ = "
]


[tool.coverage.html]
directory = "coverage_html_report"
